# task-api/server.js
const express = require("express");
const app = express();
const PORT = 3000;
const swaggerUi = require("swagger-ui-express");
const swaggerDocument = require("./openapi.json");

app.use(express.json());

app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));

let tasks = [
    {
        id:1,
        title:"Learn Express",
        done:false
    },
    {
        id:2,
        title:"Build CRUD API",
        done:false
    },
    {
        id:3,
        title:"Read Swagger Docs",
        done:true
    }
];

app.get("/tasks", (req,res)=>{
    res.json(tasks);
});

app.get("/tasks/:id",(req,res)=>{

    const id = Number(req.params.id);

    const task = tasks.find(t=>t.id===id);

    if(!task){
        return res.status(404).json({
            error:`Task ${id} not found`
        });
    }

    res.json(task);

});

app.get("/", (req, res) => {
    res.json({
        name: "Task API",
        version: "1.0",
        endpoints: ["/tasks"]
    });
});

app.post("/tasks",(req,res)=>{

    const {title}=req.body;

    if(!title || title.trim()===""){
        return res.status(400).json({
            error:"Title is required"
        });
    }

    const newTask={
        id:tasks.length ? tasks[tasks.length-1].id+1 : 1,
        title,
        done:false
    };

    tasks.push(newTask);

    res.status(201).json(newTask);

});
app.put("/tasks/:id",(req,res)=>{

    const id=Number(req.params.id);

    const task=tasks.find(t=>t.id===id);

    if(!task){
        return res.status(404).json({
            error:`Task ${id} not found`
        });
    }

    const {title,done}=req.body;

    if(title!==undefined){

        if(title.trim()===""){
            return res.status(400).json({
                error:"Title cannot be empty"
            });
        }

        task.title=title;
    }

    if(done!==undefined){
        task.done=done;
    }

    res.json(task);

});
app.delete("/tasks/:id",(req,res)=>{

    const id=Number(req.params.id);

    const index=tasks.findIndex(t=>t.id===id);

    if(index===-1){
        return res.status(404).json({
            error:`Task ${id} not found`
        });
    }

    tasks.splice(index,1);

    res.status(204).send();

});
app.get("/health", (req, res) => {
    res.json({
        status: "ok"
    });
});

app.listen(PORT, () => {
    console.log(`Server running on http://localhost:${PORT}`);
});